feat: organize folders
